backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Iris_duyendt <duyendt@codenet.jp>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.11.1"
fastapi = "^0.98.0"
uvicorn = "^0.22.0"
peewee = "^3.16.2"
pydantic = "^1.10.9"
psycopg2-binary = "^2.9.6"
mangum = "^0.17.0"
setuptools = "^68.0.0"
python-dotenv = "^1.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-i18n = {extras = ["yaml"], version = "^0.3.9"}
mecab-python3 = "^1.0.6"
jaconv = "^0.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
